Python script to sync lights to music
Uploading an old script used to sync the lights connected to an arduino board on serial to the beat of the song I currently listen to on Spotify. This is done using the spotify api. One must create a dev client with spotify and add the client id and secret.